rename modules since they no longer contain clients
